Cleanup
just erased some unused thing and moved tests to utils
