--- app/src/main/java/io/ohalloran/crypto/it/mobistego/utils/Constants.java
@@ -1,35 +1,0 @@
-package io.ohalloran.crypto.it.mobistego.utils;
-
-/**
- * Created by paspao on 28/01/15.
- * Copyright (C) 2015  Pasquale Paola
- * <p/>
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 2.1 of the License, or (at your option) any later version.
- * <p/>
- * This library is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- * <p/>
- * You should have received a copy of the GNU Lesser General Public
- * License along with this library; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
- */
-public class Constants {
-
-    public static final String CHOOSEN_IMAGE = "CHOOSEN_IMAGE";
-    public static final String EXT_DIR = "MobiStego";
-    public static final String FILE_PNG_EXT = ".png";
-    public static final String FILE_TXT_EXT = ".txt";
-    public static final String NO_NAME = "NO_NAME";
-    public static final String CONTAINER = "MYCONTAINER";
-
-    public static final int SELECT_PHOTO = 1;
-    public static final int REQUEST_IMAGE_CAPTURE = 2;
-    public static final int BACK_COMPOSE = 3;
-    public static final int SELECT_PHOTO_DECODE = 4;
-
-}
